Refactor into separate pieces, add Factory.
- [x] __to_dict__ -> to_dict
- [x] __from_dict__ -> from_dict
- [x] separate unionable from IO capabilities
- [x] separate AbstractPlottable from AbstractReport et. al.
- [x] move AbstractPlottable to Abstract.py
- [x] move AbstractReport (and subclasses) to Report.py and import ReportData.py
- [x] move AbstractReportData (and subclasses) to ReportData.py
- [x] move ReportSet (and subclasses) to ReportData.py
- [x] make sure to_dict is used over __to_dict__ (except in definitions)
- [x] make sure from_dict is used over __from_dict__ (except in definitions)
- [x] remove 'version=' from input arguments in __init__ functions
- [x] make AbstractReport itself a subclass of AbstractReportData
- [x] add a Factory dict for ReportData objects
- [x] add a Factory dict for Report objects
