+++ geco_stat/Abstract.py
@@ -6,25 +6,122 @@ import abc
 import numpy as np      # >=1.10.4
 from geco_stat._version import __version__
 from geco_stat._constants import __default_bitrate__
+from geco_stat.Exceptions import VersionException
 
-# TODO: Implement.
+# You can store classes in the factory for later recovery, but you have to
+# manually add classes to the factory after defining them.
+class Factory(object):
+    """
+    A factory class for mapping the names of AbstractDictRepresentable subclass
+    names to their corresponding subclasses. Classes can be added to the
+    factory using the ``add_class`` method and recovered using the
+    ``get_class`` method, which is useful for, e.g., reconstructing an object
+    from a dictionary when the class of that object is specified as a string.
+
+    For example, if ``d`` is such a dictionary, then
+
+        cls = geco_stat.Factory.get_class(d['class'])
+
+    can be used to get (what is presumably) the original class associated with
+    the instance represented by that dictionary. The rest of the data in that
+    dictionary should then be sufficient (along with the class itself) to
+    reconstruct the original object. At the time of writing, this is the
+    approach favored by the ``__from_dict__`` private methods used by
+    ``from_dict``.
+
+    You can store classes in the factory for later recovery, but you have to
+    manually add classes to the factory after defining them. This approach has
+    the advantage of facilitating runtime modifications for prototyping or
+    advanced munging/scripting purposes; simply defining a new class and
+    adding it to the factory should be enough to give it all of the features of
+    built-in geco_stat classes.
+    """
+
+    _factory = dict()
+
+    @staticmethod
+    def add_class(cls, name=None):
+        """
+        Store a class in the factory for later retrieval.
+
+        >>> geco_stat.Factory.add_class(object, 'object')
+        """
+        if name is None:
+            name = cls.__name__
+        Factory._factory[name] = cls 
+
+    @staticmethod
+    def get_class(clsname):
+        """
+        Retrieve a class by name:
+
+        >>> geco_stat.Factory.get_class('object')
+        object
+        """
+        return Factory._factory[clsname]
+
+    @staticmethod
+    def classes():
+        """
+        List all classes currently in the factory.
+
+        >>> geco_stat.Factory.classes()
+        [object]
+        """
+        return Factory._factory.values()
+
+    @staticmethod
+    def classnames():
+        """
+        List the names of all classes currently in the factory.
+
+        >>> geco_stat.Factory.classnames()
+        ['object']
+        """
+        return Factory._factory.keys()
+
+    @staticmethod
+    def items():
+        """
+        Return a list of key-value pair tuples containing the names
+        and classes for each class stored in the factory.
+
+        >>> geco_stat.Factory.items()
+        [('object', object)]
+        """
+        return Factory._factory.items()
+
+    @staticmethod
+    def del_class(clsname):
+        """
+        Remove a class from the factory.
+
+        >>> geco_stat.Factory.del_class('object')
+        """
+        del Factory._factory[clsname]
+
 class AbstractTimeIntervalSet(object):
     """
     The AbstractTimeIntervalSet defines what objects can specify time intervals
     over which data might be collected. In practice, it should only have one
     implementation; the abstract class serves mainly to keep the type graph
     acyclic and restrictive.
+
+    This abstract class is meant for organizational/conceptual purposes and
+    does not specify an interface or implement any methods.
     """
     __metaclass__  = abc.ABCMeta
     __version__ = __version__
 
-# TODO: Implement.
 class AbstractTimeSeries(object):
     """
     The AbstractTimeSeries provides a way of defining the category of objects
     which can serve as valid sources of information for geco_stat. In
     practice, it should only have one implementation; the abstract class serves
     mainly to keep the type graph acyclic.
+
+    This abstract class is meant for organizational/conceptual purposes and
+    does not specify an interface or implement any methods.
     """
     __metaclass__  = abc.ABCMeta
     __version__ = __version__
@@ -78,9 +175,8 @@ class AbstractNonIntersectingUnionable(AbstractTimeSeriesDerivable):
     __metaclass__  = abc.ABCMeta
     __version__ = __version__
 
-    # FIXME: This is a hack. This should be an abstract class, but most
-    # subclasses have implemented the private methods, so this is a temporary
-    # kludge.
+    # Use a template method approach to always make sure that these items
+    # are unionable before proceeding.
     def union(self, other):
         """
         Combine these two instances to cover a larger span of time. Like the
@@ -120,19 +216,21 @@ class AbstractNonIntersectingUnionable(AbstractTimeSeriesDerivable):
         """
         return self.union(other)
 
-    # TODO Decide what to do about these consistency checks. Maybe implement
-    # a very simple top-level check that can be overridden or extended as
-    # needed, but only if it is considered important; this should not distract.
-    # In any case, the private method name is undesirable and should be
+    # TODO The private method name is undesirable and should be
     # replaced with something better. Not deprecated -- for now.
-    # ALMOST deprecated...
     @abc.abstractmethod
     def _assert_self_consistent(self):
         """
         Make sure this instance is self-consistent.
         """
 
-class AbstractDictRepresentable(AbstractNonIntersectingUnionable):
+    @abc.abstractmethod
+    def _assert_unionable(self, other):
+        """
+        Make sure these two instances can be unioned.
+        """
+
+class AbstractDictRepresentable(object):
     """
     Subclasses of AbstractDictRepresentable contain information that can be
     represented as a **PARTICULAR** type of dictionary. To be specific, they
@@ -150,9 +248,6 @@ class AbstractDictRepresentable(AbstractNonIntersectingUnionable):
     __metaclass__  = abc.ABCMeta
     __version__ = __version__
 
-    # FIXME: This is a hack. This should be an abstract class, but most
-    # subclasses have implemented the private methods, so this is a temporary
-    # kludge.
     def to_dict(self):
         """
         Return a dictionary whose elements consist of strings, ints, lists, or
@@ -162,17 +257,29 @@ class AbstractDictRepresentable(AbstractNonIntersectingUnionable):
         the dictionary's contents. This is an implementation method used to
         store data in HDF5.
         """
-        return self.__to_dict__()
+        # Unless overridden, this method will always check version and class
+        # information, with class-specific treatment of the dictionary being
+        # handled by a template method, __to_dict__.
+        d = self.__to_dict__()
+        d['version'] = self.__version__
+        d['class'] = type(self).__name__
+        return d
 
-    # FIXME: This is a hack. This should be an abstract class, but most
-    # subclasses have implemented the private methods, so this is a temporary
-    # kludge.
-    def from_dict(cls, d):
+    @staticmethod
+    def from_dict(d):
         """
         Construct an instance of this class using a dictionary of the form
-        output by self.__to_dict__. Should generally be a class method.
+        output by self.to_dict. Should generally be a class method.
         """
-        return cls.__from_dict__t(d)
+        # Unless overridden, this method will always check version and class
+        # information, with class-specific treatment of the dictionary being
+        # handled by a template method, __from_dict__.
+        cls = Factory.get_class(d['class'])
+        if d['version'] != cls.__version__:
+            raise VersionException("Tried constructing an instance of %s "
+                                   "using old version %s"
+                                   % (cls.__name__, d['version']))
+        return cls.__from_dict__(d)
 
     def clone(self):
         """
@@ -184,21 +291,26 @@ class AbstractDictRepresentable(AbstractNonIntersectingUnionable):
         # If this implementation makes you nervous (for anything besides
         # performance reasons, then you are not properly implementing the
         # to_dict and from_dict methods.
-        return self.__from_dict__(self.__to_dict__())
+        return self.from_dict(self.to_dict())
 
-    # This interface should become "to_dict", since it will likely be
-    # useful for users.
-    # FIXME deprecated
     @abc.abstractmethod
     def __to_dict__(self):
-        pass
+        """
+        Return a dictionary whose elements consist of strings, ints, lists, or
+        numpy.ndarray objects, or of other dicts whose contents follow this
+        pattern recursively. This dictionary must wholly represent the data in
+        this object, so that this object may be totally reconstructed using
+        the dictionary's contents. This is an implementation method used to
+        store data in HDF5.
+        """
 
-    # This interface should become "from_dict", since it will likely be
-    # useful for users.
-    # FIXME deprecated
+    # MUST BE A CLASS METHOD.
     @abc.abstractmethod
     def __from_dict__(cls, d):
-        pass
+        """
+        Construct an instance of this class using a dictionary of the form
+        output by self.to_dict. Should generally be a class method.
+        """
 
 class HDF5_IO(AbstractDictRepresentable):
     """
@@ -210,7 +322,7 @@ class HDF5_IO(AbstractDictRepresentable):
 
     def save_hdf5(self, filename):
         """Save this instance to an hdf5 file."""
-        self.__save_dict_to_hdf5__(self.__to_dict__(), filename)
+        self.__save_dict_to_hdf5__(self.to_dict(), filename)
 
     @classmethod
     def load_hdf5(cls, filename):
@@ -218,7 +330,7 @@ class HDF5_IO(AbstractDictRepresentable):
         # TODO: Make this a AbstractReport-defined staticmethod that only
         # needs class information that is already provided in the saved
         # dictionary.
-        return cls.__from_dict__(cls.__load_dict_from_hdf5__(filename))
+        return cls.from_dict(cls.__load_dict_from_hdf5__(filename))
 
     @classmethod
     def __save_dict_to_hdf5__(cls, dic, filename):
@@ -226,7 +338,7 @@ class HDF5_IO(AbstractDictRepresentable):
         Save a dictionary whose contents are only strings, np.float64,
         np.int64, np.ndarray, and other dictionaries following this structure
         to an HDF5 file. These are the sorts of dictionaries that are meant
-        to be produced by the AbstractReport__to_dict__() method. The saved
+        to be produced by the AbstractDictRepresentable.to_dict() method. The saved
         dictionary can then be loaded using __load_dict_to_hdf5__(), and the
         contents of the loaded dictionary will be the same as those of the
         original.
@@ -280,8 +392,8 @@ class HDF5_IO(AbstractDictRepresentable):
         Load a dictionary whose contents are only strings, floats, ints,
         numpy arrays, and other dictionaries following this structure
         from an HDF5 file. These dictionaries can then be used to reconstruct
-        AbstractReport subclass instances using the
-        AbstractReport.__from_dict__() method.
+        AbstractDictRepresentable subclass instances using the
+        AbstractDictRepresentable.from_dict() method.
         """
         with h5py.File(filename, 'r') as h5file:
             return cls.__recursively_load_dict_contents_from_group__(h5file,
@@ -302,17 +414,26 @@ class HDF5_IO(AbstractDictRepresentable):
                     h5file, path + key + '/')
         return ans
 
-# Rename AbstractReport; it should be called something more
-# usage-agnostic and feature-descriptive.
-# FIXME deprecated
-class AbstractReport(HDF5_IO):
-    "Abstract interface used by all geco_statistics classes"
+class AbstractPlottable(object):
+    """
+    An interface for generating matplotlib figures that can be used in
+    visualizing data.
+    """
     __metaclass__  = abc.ABCMeta
-    __version__ = __version__
 
-    # This should be handled through ducktyping.
-    # FIXME deprecated
     @abc.abstractmethod
-    def _assert_unionable(self, other):
-        "Make sure these two instances can be unioned."
+    def plot(self):
+        """
+        Create some sort of visualization for the information content of this
+        object. Container objects should use this to
+        recursively call plotting functions in their constituents and generate
+        summary plots representing all their information.
+        """
+
+    @abc.abstractmethod
+    def summary(self):
+        """
+        Create some sort of verbose, human-readable text summary for the
+        information content of this object. Should return a string.
+        """
