+++ geco_stat/_version.py
@@ -1,2 +1,2 @@
 __version__ = '0.2'
-__release__ = '0.2.10'
+__release__ = '0.2.11'
